# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "같이살림 API",
    "version": "1.0.0",
    "description": "같이살림 API 명세서입니다."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "인증"
        ],
        "summary": "내 정보 조회",
        "description": "내 정보 조회",
        "operationId": "getAuthMe",
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "userId",
                        "name",
                        "nickname",
                        "email",
                        "profileImage"
                      ],
                      "properties": {
                        "userId": {
                          "type": "integer",
                          "example": 1
                        },
                        "name": {
                          "type": "string",
                          "example": "홍길동"
                        },
                        "nickname": {
                          "type": "string",
                          "example": "길동"
                        },
                        "email": {
                          "type": "string",
                          "format": "email",
                          "example": "user@example.com"
                        },
                        "profileImage": {
                          "type": "string",
                          "format": "uri",
                          "nullable": true,
                          "example": "https://example.com/profile.png"
                        }
                      }
                    },
                    "error": {
                      "nullable": true
                    }
                  }
                },
                "example": {
                  "statusCode": 200,
                  "data": {
                    "userId": 1,
                    "name": "홍길동",
                    "nickname": "길동",
                    "email": "user@example.com",
                    "profileImage": "https://example.com/profile.png"
                  },
                  "error": null
                }
              }
            }
          },
          "400": {
            "description": "실패",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "data": {
                      "nullable": true
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "errors": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "value": {
                                "type": "string"
                              },
                              "reason": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                },
                "example": {
                  "statusCode": 400,
                  "data": null,
                  "error": {
                    "code": "COMMON_INVALID_PARAMETER",
                    "message": "요청 파라미터가 잘못되었습니다.",
                    "errors": [
                      {
                        "field": "nickname",
                        "value": "홍길동!",
                        "reason": "특수문자는 포함할 수 없습니다."
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "실패",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "data": {
                      "nullable": true
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "errors": {
                          "type": "array",
                          "items": {}
                        }
                      }
                    }
                  }
                },
                "example": {
                  "statusCode": 401,
                  "data": null,
                  "error": {
                    "code": "AUTH_UNAUTHORIZED",
                    "message": "인증이 필요합니다.",
                    "errors": []
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      },
      "delete": {
        "tags": [
          "인증"
        ],
        "summary": "회원 탈퇴",
        "description": "회원 탈퇴",
        "operationId": "deleteAuthMe",
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "userId",
                        "deleted"
                      ],
                      "properties": {
                        "userId": {
                          "type": "integer",
                          "example": 1
                        },
                        "deleted": {
                          "type": "boolean",
                          "enum": [
                            true
                          ],
                          "example": true
                        }
                      }
                    },
                    "error": {
                      "nullable": true
                    }
                  }
                },
                "example": {
                  "statusCode": 200,
                  "data": {
                    "userId": 1,
                    "deleted": true
                  },
                  "error": null
                }
              }
            }
          },
          "400": {
            "description": "실패",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "data": {
                      "nullable": true
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "errors": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "value": {
                                "type": "string"
                              },
                              "reason": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                },
                "example": {
                  "statusCode": 400,
                  "data": null,
                  "error": {
                    "code": "COMMON_INVALID_PARAMETER",
                    "message": "요청 파라미터가 잘못되었습니다.",
                    "errors": [
                      {
                        "field": "nickname",
                        "value": "홍길동!",
                        "reason": "특수문자는 포함할 수 없습니다."
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "실패",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "data": {
                      "nullable": true
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "errors": {
                          "type": "array",
                          "items": {}
                        }
                      }
                    }
                  }
                },
                "example": {
                  "statusCode": 401,
                  "data": null,
                  "error": {
                    "code": "AUTH_UNAUTHORIZED",
                    "message": "인증이 필요합니다.",
                    "errors": []
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "502": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/auth/profile": {
      "patch": {
        "tags": [
          "인증"
        ],
        "summary": "프로필 수정",
        "description": "이름, 닉네임 또는 프로필 이미지를 수정합니다.",
        "operationId": "patchAuthProfile",
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "userId",
                        "name",
                        "nickname",
                        "email",
                        "profileImage"
                      ],
                      "properties": {
                        "userId": {
                          "type": "integer",
                          "example": 1
                        },
                        "name": {
                          "type": "string",
                          "example": "홍길동"
                        },
                        "nickname": {
                          "type": "string",
                          "example": "길동"
                        },
                        "email": {
                          "type": "string",
                          "format": "email",
                          "example": "user@example.com"
                        },
                        "profileImage": {
                          "type": "string",
                          "format": "uri",
                          "nullable": true,
                          "example": null
                        }
                      }
                    },
                    "error": {
                      "nullable": true
                    }
                  }
                },
                "example": {
                  "statusCode": 200,
                  "data": {
                    "userId": 1,
                    "name": "홍길동",
                    "nickname": "길동",
                    "email": "user@example.com",
                    "profileImage": null
                  },
                  "error": null
                }
              }
            }
          },
          "400": {
            "description": "실패",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "data": {
                      "nullable": true
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "errors": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "value": {
                                "type": "string"
                              },
                              "reason": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                },
                "example": {
                  "statusCode": 400,
                  "data": null,
                  "error": {
                    "code": "COMMON_INVALID_PARAMETER",
                    "message": "요청 파라미터가 잘못되었습니다.",
                    "errors": [
                      {
                        "field": "nickname",
                        "value": "홍길동!",
                        "reason": "특수문자는 포함할 수 없습니다."
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "실패",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "data": {
                      "nullable": true
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "errors": {
                          "type": "array",
                          "items": {}
                        }
                      }
                    }
                  }
                },
                "example": {
                  "statusCode": 401,
                  "data": null,
                  "error": {
                    "code": "AUTH_UNAUTHORIZED",
                    "message": "인증이 필요합니다.",
                    "errors": []
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "이름 (1~30자)",
                    "minLength": 1,
                    "maxLength": 30,
                    "example": "홍길동"
                  },
                  "nickname": {
                    "type": "string",
                    "description": "닉네임 (2~10자, 특수문자 불가)",
                    "minLength": 2,
                    "maxLength": 10,
                    "pattern": "^[가-힣a-zA-Z0-9]+$",
                    "example": "길동"
                  },
                  "profileImage": {
                    "type": "string",
                    "format": "uri",
                    "maxLength": 512,
                    "nullable": true,
                    "description": "프로필 이미지 URL. null이면 기존 이미지를 제거합니다.",
                    "example": null
                  }
                },
                "minProperties": 1
              },
              "example": {
                "name": "홍길동",
                "nickname": "길동",
                "profileImage": null
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/profile-image/upload-url": {
      "post": {
        "tags": [
          "인증"
        ],
        "summary": "프로필 이미지 업로드 URL 발급",
        "description": "최대 5MB의 JPEG, PNG 또는 WebP 파일을 업로드할 수 있는 5분 유효 S3 Presigned POST를 발급합니다.",
        "operationId": "postAuthProfileImageUploadUrl",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProfileImageUpload"
              },
              "example": {
                "contentType": "image/jpeg",
                "fileSize": 824123
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "프로필 이미지 업로드 URL 발급 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "data",
                    "error"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 201
                    },
                    "data": {
                      "$ref": "#/components/schemas/ProfileImageUploadResponse"
                    },
                    "error": {
                      "nullable": true,
                      "example": null
                    }
                  }
                },
                "example": {
                  "statusCode": 201,
                  "data": {
                    "uploadMethod": "POST",
                    "uploadUrl": "https://profile-images.s3.ap-northeast-2.amazonaws.com",
                    "fields": {
                      "key": "develop/profiles/7/550e8400-e29b-41d4-a716-446655440000.jpg",
                      "Content-Type": "image/jpeg",
                      "policy": "base64-policy",
                      "x-amz-signature": "signature"
                    },
                    "objectKey": "develop/profiles/7/550e8400-e29b-41d4-a716-446655440000.jpg",
                    "profileImageUrl": "https://example.cloudfront.net/develop/profiles/7/550e8400-e29b-41d4-a716-446655440000.jpg",
                    "expiresAt": "2026-07-28T12:05:00.000Z"
                  },
                  "error": null
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Cognito access token을 Authorization: Bearer 헤더로 전달합니다."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer"
          },
          "data": {
            "nullable": true
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "errors": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "CreateProfileImageUpload": {
        "type": "object",
        "required": [
          "contentType",
          "fileSize"
        ],
        "properties": {
          "contentType": {
            "type": "string",
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp"
            ],
            "example": "image/jpeg"
          },
          "fileSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5242880,
            "example": 824123
          }
        }
      },
      "ProfileImageUploadResponse": {
        "type": "object",
        "required": [
          "uploadMethod",
          "uploadUrl",
          "fields",
          "objectKey",
          "profileImageUrl",
          "expiresAt"
        ],
        "properties": {
          "uploadMethod": {
            "type": "string",
            "enum": [
              "POST"
            ],
            "example": "POST"
          },
          "uploadUrl": {
            "type": "string",
            "format": "uri",
            "example": "https://profile-images.s3.ap-northeast-2.amazonaws.com"
          },
          "fields": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "objectKey": {
            "type": "string",
            "example": "develop/profiles/7/550e8400-e29b-41d4-a716-446655440000.jpg"
          },
          "profileImageUrl": {
            "type": "string",
            "format": "uri",
            "example": "https://example.cloudfront.net/develop/profiles/7/550e8400-e29b-41d4-a716-446655440000.jpg"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-07-28T12:05:00.000Z"
          }
        }
      },
...
    },
    "responses": {
      "ErrorResponse": {
        "description": "오류 응답",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}